pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mango-brain"
version = "3.0.0"
description = "Persistent associative memory system for Claude Code — complete development workflow"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Federico Anastasi", email = "federico.anastasi@outlook.com"},
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "mcp>=1.0",
    "sentence-transformers>=2.5",
    "torch>=2.2",
    "aiosqlite>=0.19",
    "numpy>=1.26",
    "scipy>=1.12",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "tiktoken>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
mango-brain = "server.cli:main"

[tool.setuptools.packages.find]
include = ["server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
